refactor(schemas): migrate Phase 16 endpoint schemas — 6 files, 24 classes + 1 enum (#6042)
- triggers.py: migrate TriggerCreateRequest, TriggerCreateResponse,
  TriggerListResponse, FireTriggerRequest → schemas_workflows.py
- service_messages.py: migrate ServiceMessageResponse, LatestMessagesResponse,
  SingleMessageResponse, CorrelationChainResponse → schemas_system.py
- project_state.py: migrate PhaseStatus, ProjectStatus, ValidationResultModel,
  PhaseValidationModel → schemas_system.py
- services.py: migrate ServiceStatus, SystemInfo, VMStatus, ServicesResponse
  → schemas_system.py
- research_browser.py: migrate ResearchRequest→BrowserResearchRequest,
  SessionAction, NavigationRequest, CreateChatBrowserRequest → schemas_workflows.py
- structured_thinking_mcp.py: migrate MCPTool→StructuredThinkingMCPTool,
  ProcessThoughtRequest (with helper methods), GenerateSummaryRequest,
  ClearHistoryRequest → schemas_system.py; also migrate ThinkingStage enum

Renames to avoid generic-name collisions:
- ResearchRequest → BrowserResearchRequest (collides with knowledge ResearchRequest)
- MCPTool → StructuredThinkingMCPTool (3rd MCPTool variant)

schemas_workflows.py: imports TriggerType from services.trigger_service
schemas_system.py: gains Enum import for ThinkingStage

Bug fix: research_browser.py line 81 referenced bare `research_browser_manager`
(undefined) instead of the imported `get_research_browser_manager`. Pre-existing
typo, fixed inline since it was blocking flake8 F821.

Cleanup: removed dead-code FireTriggerRequest import in triggers.py and
NetworkConstants/Optional in research_browser.py.
