+++ autobot-backend/api/service_messages.py
@@ -13,13 +13,18 @@ Endpoints:
 """
 
 import logging
-from typing import List, Optional
+from typing import Optional
 
 from fastapi import APIRouter, Depends, HTTPException, Query, Request
-from pydantic import BaseModel, Field
 
 from auth_middleware import get_auth_middleware
 from autobot_shared.message_bus import get_message_bus
+from api.schemas_system import (
+    CorrelationChainResponse,
+    LatestMessagesResponse,
+    ServiceMessageResponse,
+    SingleMessageResponse,
+)
 from utils.catalog_http_exceptions import raise_auth_error, raise_server_error
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 
@@ -31,43 +36,6 @@ router = APIRouter(prefix="/service-messages", tags=["service-messages"])
 # ------------------------------------------------------------------
 # Response models
 # ------------------------------------------------------------------
-
-
-class ServiceMessageResponse(BaseModel):
-    """Single serialised ServiceMessage."""
-
-    msg_id: str
-    ts: str
-    sender: str
-    receiver: str
-    msg_type: str
-    content: str
-    correlation_id: str
-    meta: dict = Field(default_factory=dict)
-
-
-class LatestMessagesResponse(BaseModel):
-    """Response for GET /latest."""
-
-    success: bool
-    count: int
-    messages: List[ServiceMessageResponse]
-
-
-class SingleMessageResponse(BaseModel):
-    """Response for GET /{msg_id}."""
-
-    success: bool
-    message: Optional[ServiceMessageResponse] = None
-
-
-class CorrelationChainResponse(BaseModel):
-    """Response for GET /chain/{correlation_id}."""
-
-    success: bool
-    correlation_id: str
-    count: int
-    messages: List[ServiceMessageResponse]
 
 
 # ------------------------------------------------------------------
